Move public SSE facade to HTTP package

Copilot-Session: ff0a367c-664e-4de0-902b-ddaae2c86546
