+++ sdk/core/azure-core/src/main/java/com/azure/core/http/ServerSentEventStreams.java
@@ -1,10 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.core.http.rest;
+package com.azure.core.http;
 
-import com.azure.core.http.ServerSentEvent;
-import com.azure.core.http.ServerSentEventListener;
+import com.azure.core.http.rest.Response;
 import com.azure.core.implementation.util.ServerSentEventStream;
 import com.azure.core.util.BinaryData;
 import reactor.core.publisher.Flux;
