+++ sdk/core/azure-core/src/test/java/com/azure/core/implementation/util/ServerSentEventStreamTests.java
@@ -7,8 +7,8 @@ import com.azure.core.http.HttpHeaderName;
 import com.azure.core.http.HttpHeaders;
 import com.azure.core.http.ServerSentEvent;
 import com.azure.core.http.ServerSentEventListener;
+import com.azure.core.http.ServerSentEventStreams;
 import com.azure.core.http.rest.ResponseBase;
-import com.azure.core.http.rest.ServerSentEventStreams;
 import com.azure.core.util.BinaryData;
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
